requirements.txt
django
openai
python-dotenv
django-allauth